[Phase 2] Implement optimizer for wms (#266)
* Change to use enum to express object from both API model and DB model

* Implement logic to store 3rd tokens

* Implement CURD of topology APIs

* Implement topology module

* Implement new tables to support picking phase

* Implement API to assign shelves to nodes

* Implement API to create new paths

* Implement state machine to manage picking flow

* Fix wrong method

* Implement CURD for picking wave feature

* Draft implement of state machine to manage changing flow in both Route, Plan and Sale

* Move from i32 to i64

* Implement ShowflakeID

* Correct sale event

* Implement CRUD of picking

* Split get item into two APIs to avoid coupling

* Implement new APIs to support change picking route status

+++ pkgs/backend/src/api/wms/mod.rs
@@ -6,7 +6,7 @@ use std::future::{ready, Ready};
 
 #[derive(Debug)]
 pub struct WmsHeaders {
-    tenant_id: i32,
+    tenant_id: i64,
     is_guess: bool,
 }
 
@@ -29,7 +29,7 @@ impl FromRequest for WmsHeaders {
         };
         let tenant_id = match headers.get("X-Tenant-Id") {
             Some(value) => match value.to_str() {
-                Ok(str_val) => match str_val.parse::<i32>() {
+                Ok(str_val) => match str_val.parse::<i64>() {
                     Ok(parsed) => parsed,
                     Err(_) => {
                         return ready(Err(ErrorBadRequest(
